+++ developer/src/common/web/utils/index.ts
@@ -1,1 +1,2 @@
 export { validateMITLicense } from './src/validate-mit-license.js';
+export { KeymanSentry } from './src/KeymanSentry.js';
